+++ openeo/rest/datacube.py
@@ -300,7 +300,7 @@ class DataCube(_ProcessGraphAbstraction):
                 unsupported_properties = set(properties.keys()).difference(supported_properties)
                 if unsupported_properties:
                     warnings.warn(
-                        f"Property filtering with unsupported properties according to collection/STAC metadata: {unsupported_properties} (supported: {supported_properties}).",
+                        f"Property filtering with properties not listed in collection/STAC metadata: {list(unsupported_properties)} (supported: {list(supported_properties)}).",
                         stacklevel=3,
                     )
             properties = {
@@ -2456,7 +2456,7 @@ class DataCube(_ProcessGraphAbstraction):
         :param additional: (optional) additional (top-level) properties to set in the request body
         :param job_options: (optional) dictionary of job options to pass to the backend
             (under top-level property "job_options")
-        :param on_response_headers: (optional) callback to handle/show the response headers
+        :param on_response_headers: (optional) callback to handle (e.g. :py:func:`print`) the response headers.
 
         :return: if ``outputfile`` was not specified:
             a :py:class:`bytes` object containing the raw data.
@@ -3069,6 +3069,34 @@ class DataCube(_ProcessGraphAbstraction):
                 dimension=dimension,
                 target_dimensions=target_dimensions,
                 label_separator=label_separator,
+            ),
+        )
+
+    @openeo_process
+    def aspect(self) -> DataCube:
+        """
+        Converts every band in the datacube to a band with suffix "_aspect" containing the aspect.
+
+        :return: A data cube with the same dimensions but containing the computed aspect values.
+        """
+        return self.process(
+            process_id="aspect",
+            arguments=dict_no_none(
+                data=THIS
+            ),
+        )
+
+    @openeo_process
+    def slope(self) -> DataCube:
+        """
+        Converts every band in the datacube to a band with suffix "_slope" containing the slope.
+
+        :return: A data cube with the same dimensions but containing the computed slope values.
+        """
+        return self.process(
+            process_id="slope",
+            arguments=dict_no_none(
+                data=THIS
             ),
         )