+++ openeo/rest/vectorcube.py
@@ -268,7 +268,7 @@ class VectorCube(_ProcessGraphAbstraction):
         :param additional: (optional) additional (top-level) properties to set in the request body
         :param job_options: (optional) dictionary of job options to pass to the backend
             (under top-level property "job_options")
-        :param on_response_headers: (optional) callback to handle/show the response headers
+        :param on_response_headers: (optional) callback to handle (e.g. :py:func:`print`) the response headers.
 
         :return: if ``outputfile`` was not specified:
             a :py:class:`bytes` object containing the raw data.
@@ -292,7 +292,12 @@ class VectorCube(_ProcessGraphAbstraction):
         else:
             res = self
         return self._connection.download(
-            res.flat_graph(), outputfile=outputfile, validate=validate, additional=additional, job_options=job_options
+            res.flat_graph(),
+            outputfile=outputfile,
+            validate=validate,
+            additional=additional,
+            job_options=job_options,
+            on_response_headers=on_response_headers,
         )
 
     def execute_batch(